+++ DevToolkt/reactive/src/commonMain/kotlin/dev/toolkt/reactive/event_stream/BoundListener.kt
@@ -1,0 +1,9 @@
+package dev.toolkt.reactive.event_stream
+
+import dev.toolkt.reactive.Subscription
+
+interface BoundListener {
+    fun listen(): Subscription
+
+    fun listenWeak(): Subscription
+}
